check if file exists

+++ quasimodo/assertion_generation/openie_reader.py
@@ -57,7 +57,8 @@ class OpenIEReader(object):
                 try:
                     with open(filename_no_found, "a") as f:
                         f.write(sentence + "\n")
-                        os.remove(CACHE_OPENIE_READER)
+                        if os.path.isfile(CACHE_OPENIE_READER):
+                            os.remove(CACHE_OPENIE_READER)
                         break
                 except Exception as e:
                     logging.info("Error while writting the sentence in openie_reader")
